Automatically exclude connection /disconnection joins of systems defined in the GUI from binding with CrestronMobile
The idea is not to clear these joins when CrestronMobile resets all the joins on connection. We automatically scan all systems defined in the GUI, and if they have a connection and/or disconnection join that is not set to 0, add it to the exclusion list to prevent CrestronMobile from even touching it.